+++ tests/e2e/assertions_test.go
@@ -16,6 +16,7 @@ func Test_E2E_Assertion_NoCalls(t *testing.T) {
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
 		true,
+		nil,
 	)
 
 	defer killMock()
@@ -44,6 +45,7 @@ func Test_E2E_Assertion_BasicAssertion_WithValidationErrors(t *testing.T) {
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
 		true,
+		nil,
 	)
 	defer killMock()
 
@@ -79,6 +81,7 @@ func Test_E2E_Assertion_WithNth(t *testing.T) {
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
 		true,
+		nil,
 	)
 	defer killMock()
 
@@ -111,6 +114,7 @@ func Test_E2E_Assertion_WithNth_Failing(t *testing.T) {
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
 		true,
+		nil,
 	)
 	defer killMock()
 
@@ -150,6 +154,7 @@ func Test_E2E_Assertion_WithNth_OutOfRange(t *testing.T) {
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
 		true,
+		nil,
 	)
 	defer killMock()
 
@@ -185,6 +190,7 @@ func Test_E2E_Assertion_BasicAssertion_WithoutValidationErrors(t *testing.T) {
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
 		true,
+		nil,
 	)
 	defer killMock()
 
@@ -212,6 +218,7 @@ func Test_E2E_Assertion_Chaining_WithValidationErrors(t *testing.T) {
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
 		true,
+		nil,
 	)
 	defer killMock()
 
@@ -252,6 +259,7 @@ func Test_E2E_Assertion_Chaining_WithoutValidationErrors(t *testing.T) {
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
 		true,
+		nil,
 	)
 	defer killMock()
 
@@ -287,6 +295,7 @@ func Test_E2E_Assertion_MethodMatchingIsCaseInsensitive(t *testing.T) {
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
 		true,
+		nil,
 	)
 	defer killMock()