+++ tests/e2e/cors_test.go
@@ -15,6 +15,7 @@ func Test_E2E_Cors_HeadersAreSet(t *testing.T) {
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}} --cors",
 		nil,
 		true,
+		nil,
 	)
 	defer killMock()
 
@@ -34,6 +35,7 @@ func Test_E2E_Cors_HeadersAreNotSet(t *testing.T) {
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
 		true,
+		nil,
 	)
 	defer killMock()