+++ tests/e2e/reset_test.go
@@ -16,6 +16,7 @@ func Test_E2E_Resetting(t *testing.T) {
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
 		true,
+		nil,
 	)
 	defer killMock()